chore: move CustomBackdrop from style to ui folder
